remove vital display ui state
