+++ app/src/main/kotlin/edu/stanford/bdh/engagehf/bluetooth/data/models/VitalDisplayData.kt
@@ -2,18 +2,6 @@ package edu.stanford.bdh.engagehf.bluetooth.data.models
 
 import edu.stanford.bdh.engagehf.bluetooth.component.OperationStatus
 
-data class VitalDisplayUiState(
-    val bloodPressure: VitalDisplayData = VitalDisplayData(
-        title = "Blood Pressure",
-    ),
-    val heartRate: VitalDisplayData = VitalDisplayData(
-        title = "Heart Rate",
-    ),
-    val weight: VitalDisplayData = VitalDisplayData(
-        title = "Weight",
-    ),
-)
-
 data class VitalDisplayData(
     val title: String,
     val value: String? = null,
